Change the way we structure the app
