+++ api/README.md
@@ -1,0 +1,3 @@
+# API
+
+This module contains our Python based API. The project is managed with [poetry](https://python-poetry.org/).
