+++ api/tests/test_test.py
@@ -1,0 +1,2 @@
+def test_something_works():
+    assert True
