--- server/requirements.txt
@@ -1,4 +1,0 @@
-fastapi>=0.68.0,<0.69.0
-pydantic>=1.8.0,<2.0.0
-uvicorn>=0.15.0,<0.16.0
-pymongo>=4.1.1
